refactor: Change location CustomerPort to WorkspacePort
