+++ peauty-customer-api/src/main/java/com/peauty/customer/business/customer/CustomerPort.java
@@ -2,10 +2,7 @@ package com.peauty.customer.business.customer;
 
 import com.peauty.customer.business.auth.dto.SignUpCommand;
 import com.peauty.domain.customer.Customer;
-import com.peauty.domain.designer.Designer;
-import com.peauty.domain.designer.Workspace;
 
-import java.util.List;
 import java.util.Optional;
 
 public interface CustomerPort {
@@ -18,7 +15,6 @@ public interface CustomerPort {
     Customer getByCustomerIdWithoutPuppies(Long customerId);
     Customer getCustomerById(Long customerId);
 
-    List<Workspace> findAllWorkspaceByAddress(String address);
-    Designer findDesignerById(Long designerId);
+
 
 }
